@@ -3,22 +3,22 @@ const proc = require('process');
 
 var accessKey = proc.env.QINIU_ACCESS_KEY;
 var secretKey = proc.env.QINIU_SECRET_KEY;
+
+var bucket = proc.env.QINIU_TEST_BUCKET;
+
 var mac = new qiniu.auth.digest.Mac(accessKey, secretKey);
 var config = new qiniu.conf.Config();
-// config.useHttpsDomain = true;
 config.zone = qiniu.zone.Zone_z0;
+config.useHttpsDomain = 'https';
 var bucketManager = new qiniu.rs.BucketManager(mac, config);
-var bucket = proc.env.QINIU_TEST_BUCKET;
-var key = 'qiniu.mp4';
-var newType = 1; // 低频存储
-
-bucketManager.changeType(bucket, key, newType, function (err, respBody, respInfo) {
+// @param bucketName 空间名
+bucketManager.getBucketInfo(bucket, function (err, respBody, respInfo) {
     if (err) {
         console.log(err);
-    // throw err;
-    } else {
-    // 200 is success
-        console.log(respInfo.statusCode);
-        console.log(respBody);
+        throw err;
+    }
+    if (respInfo.status == 200) {
+        console.log('---respBody\n' + JSON.stringify(respBody) + '\n---');
+        console.log('---respInfo\n' + JSON.stringify(respInfo) + '\n---');
     }
 });
