@@ -2,12 +2,11 @@ const util = require('./util');
 const rpc = require('./rpc');
 const conf = require('./conf');
 const digest = require('./auth/digest');
-const zone = require('./zone.js');
 const querystring = require('querystring');
 
 exports.OperationManager = OperationManager;
 
-function OperationManager(mac, config) {
+function OperationManager (mac, config) {
     this.mac = mac || new digest.Mac();
     this.config = config || new conf.Config();
 }
@@ -21,59 +20,37 @@ function OperationManager(mac, config) {
 //                  notifyURL 回调业务服务器，通知处理结果
 //                  force     结果是否强制覆盖已有的同名文件
 // @param callbackFunc(err, respBody, respInfo) - 回调函数
-OperationManager.prototype.pfop = function(bucket, key, fops, pipeline,
+OperationManager.prototype.pfop = function (bucket, key, fops, pipeline,
     options, callbackFunc) {
     options = options || {};
-    var that = this;
-    //必须参数
+    // 必须参数
     var reqParams = {
         bucket: bucket,
         key: key,
         pipeline: pipeline,
-        fops: fops.join(';'),
+        fops: fops.join(';')
     };
 
-    //notifyURL
+    // notifyURL
     if (options.notifyURL) {
         reqParams.notifyURL = options.notifyURL;
     }
 
-    //force
+    // force
     if (options.force) {
         reqParams.force = 1;
     }
 
-    var useCache = false;
-    if (this.config.zone) {
-        if (this.config.zoneExpire == -1) {
-            useCache = true;
-        } else {
-            if (!util.isTimestampExpired(this.config.zoneExpire)) {
-                useCache = true;
-            }
+    util.prepareZone(this, this.mac.accessKey, bucket, function (err, ctx) {
+        if (err) {
+            callbackFunc(err, null, null);
+            return;
         }
-    }
-
-    if (useCache) {
-        pfopReq(this.mac, this.config, reqParams, callbackFunc);
-    } else {
-        zone.getZoneInfo(this.mac.accessKey, bucket, function(err, cZoneInfo,
-            cZoneExpire) {
-            if (err) {
-                callbackFunc(err, null, null);
-                return;
-            }
-
-            //update object
-            that.config.zone = cZoneInfo;
-            that.config.zoneExpire = cZoneExpire;
-            //pfopReq
-            pfopReq(that.mac, that.config, reqParams, callbackFunc);
-        });
-    }
+        pfopReq(ctx.mac, ctx.config, reqParams, callbackFunc);
+    });
 };
 
-function pfopReq(mac, config, reqParams, callbackFunc) {
+function pfopReq (mac, config, reqParams, callbackFunc) {
     var scheme = config.useHttpsDomain ? 'https://' : 'http://';
     var requestURI = scheme + config.zone.apiHost + '/pfop/';
     var reqBody = querystring.stringify(reqParams);
@@ -84,12 +61,12 @@ function pfopReq(mac, config, reqParams, callbackFunc) {
 // 查询持久化数据处理进度
 // @param persistentId
 // @callbackFunc(err, respBody, respInfo) - 回调函数
-OperationManager.prototype.prefop = function(persistentId, callbackFunc) {
+OperationManager.prototype.prefop = function (persistentId, callbackFunc) {
     var apiHost = 'api.qiniu.com';
-    if(this.config.zone) {
-        apiHost=this.config.zone.apiHost;
+    if (this.config.zone) {
+        apiHost = this.config.zone.apiHost;
     }
-  
+
     var scheme = this.config.useHttpsDomain ? 'https://' : 'http://';
     var requestURI = scheme + apiHost + '/status/get/prefop';
     var reqParams = {
